Prevent edit-feature-overlay from hidding.
The new "zera" library for the interactivity (github.com/CartoDB/zera) keeps propagating the events once they are fired while wax didn`t.

This causes a bug in the edit-feature overlay because it was set to be hidden on map click.

To fix this, just set a timeout in order to ignore the first click (the one which displays the overlay) and react to the next one.